update board/standard board
