+++ src/main/Board/Board.java
@@ -1,37 +1,173 @@
 package Board;
 
+import Enums.Colors;
+import Pieces.Bishop;
+import Pieces.King;
+import Pieces.Knight;
+import Pieces.Pawn;
 import Pieces.Piece;
+import Pieces.Queen;
+import Pieces.Rook;
 
-import java.util.Arrays;
+import java.util.LinkedList;
 import java.util.List;
 
-public abstract class Board {
+public class Board extends StandardBoard {
 
-    //Class variables
-    protected List<Piece> blackPieceList;
-    protected List<Piece> whitePieceList;
-    protected Tile[][] tiles;
+    private List<Piece> blackPieceList;
+    private List<Piece> whitePieceList;
 
-    //Methods
-    public List<Piece> getBlackPieceList() {
-        return blackPieceList;
-    }
-    public List<Piece> getWhitePieceList() {
-        return whitePieceList;
+    public Board(){
+        super();
+        blackPieceList = createWhitePieceList(tiles);
+        whitePieceList = createBlackPieceList(tiles);
     }
 
-    public Tile[][] getTiles() {
-        return tiles;
+    private List<Piece> createBlackPieceList(Tile[][] tiles) {
+        List<Piece> pieceList = new LinkedList<>();
+
+        pieceList.add(new Rook(Colors.BLACK, tiles[7][0]));
+        pieceList.add(new Knight(Colors.BLACK, tiles[7][1]));
+        pieceList.add(new Bishop(Colors.BLACK, tiles[7][2]));
+        pieceList.add(new Queen(Colors.BLACK, tiles[7][3]));
+        pieceList.add(new King(Colors.BLACK, tiles[7][4]));
+        pieceList.add(new Bishop(Colors.BLACK, tiles[7][5]));
+        pieceList.add(new Knight(Colors.BLACK, tiles[7][6]));
+        pieceList.add(new Rook(Colors.BLACK, tiles[7][7]));
+
+        for (int i = 0; i < 8; i++) {
+            pieceList.add(new Pawn(Colors.BLACK, tiles[6][i]));
+        }
+
+        return pieceList;
     }
 
-    protected abstract Tile[][] createTiles();
+    private List<Piece> createWhitePieceList(Tile[][] tiles) {
+        List<Piece> pieceList = new LinkedList<>();
 
-    protected abstract List<Piece> createBlackPieceList(Tile[][] tiles);
+        pieceList.add(new Rook(Colors.WHITE, tiles[0][0]));
+        pieceList.add(new Knight(Colors.WHITE, tiles[0][1]));
+        pieceList.add(new Bishop(Colors.WHITE, tiles[0][2]));
+        pieceList.add(new Queen(Colors.WHITE, tiles[0][3]));
+        pieceList.add(new King(Colors.WHITE, tiles[0][4]));
+        pieceList.add(new Bishop(Colors.WHITE, tiles[0][5]));
+        pieceList.add(new Knight(Colors.WHITE, tiles[0][6]));
+        pieceList.add(new Rook(Colors.WHITE, tiles[0][7]));
 
-    protected abstract List<Piece> createWhitePieceList(Tile[][] tiles);
+        for (int i = 0; i < 8; i++) {
+            pieceList.add(new Pawn(Colors.WHITE, tiles[1][i]));
+        }
 
-    protected abstract boolean move(Piece movedPiece, Tile destination);
+        return pieceList;
+    }
 
+    public boolean move(Piece movedPiece, Tile destination) {
+        boolean flag = false;
+        switch (movedPiece.getPieceColor()) {
+            case WHITE:
+                flag = moveWhitePiece(movedPiece, destination);
+                break;
+            case BLACK:
+                flag = moveBlackPiece(movedPiece, destination);
+                break;
+        }
+
+        return flag;
+    }
+
+    private boolean moveWhitePiece(Piece movedPiece, Tile destination) {
+        boolean flag = false;
+        if(tileEmpty(destination))
+        {
+            movedPiece.setLocation(destination);
+            flag = true;
+        }
+        else
+        {
+            if(tileHasBlackPiece(destination))
+            {
+                flag = capturePiece(movedPiece,destination);
+            }
+            else
+            {
+                //Do nothing because tile already has a white piece on it
+            }
+        }
+        return flag;
+    }
+
+    private boolean moveBlackPiece(Piece movedPiece, Tile destination) {
+        boolean flag = false;
+        if(tileEmpty(destination))
+        {
+            movedPiece.setLocation(destination);
+            flag = true;
+        }
+        else
+        {
+            if(tileHasWhitePiece(destination))
+            {
+                flag = capturePiece(movedPiece,destination);
+            }
+            else
+            {
+                //Do nothing because tile already has a white piece on it
+            }
+        }
+        return flag;
+    }
+
+    private boolean capturePiece(Piece movedPiece, Tile destination) {
+        //should be false unless moveValid
+        boolean flag = true;
+        /*
+        if(moveValid(movedPiece, destination))
+        {
+            //capture Piece
+            flag = true;
+        }
+        */
+        return flag;
+    }
+
+    private boolean tileHasBlackPiece(Tile destination) {
+        boolean flag = false;
+        for (int i = 0; i < blackPieceList.size(); i++) {
+            if(destination.equals(blackPieceList.get(i).getLocation()))
+            {
+                flag = true;
+            }
+        }
+        return flag;
+    }
+
+    private boolean tileHasWhitePiece(Tile destination) {
+        boolean flag = false;
+        for (int i = 0; i < whitePieceList.size(); i++) {
+            if(destination.equals(whitePieceList.get(i).getLocation()))
+            {
+                flag = true;
+            }
+        }
+        return flag;
+    }
+
+    private boolean tileEmpty(Tile destination) {
+        boolean flag = true;
+        for (int i = 0; i < blackPieceList.size(); i++) {
+            if(destination.equals(blackPieceList.get(i).getLocation()))
+            {
+                flag = false;
+            }
+        }
+        for (int i = 0; i < whitePieceList.size(); i++) {
+            if(destination.equals(whitePieceList.get(i).getLocation()))
+            {
+                flag = false;
+            }
+        }
+        return flag;
+    }
 
     @Override
     public boolean equals(Object o) {
@@ -42,16 +178,13 @@ public abstract class Board {
 
         if (blackPieceList != null ? !blackPieceList.equals(board.blackPieceList) : board.blackPieceList != null)
             return false;
-        if (whitePieceList != null ? !whitePieceList.equals(board.whitePieceList) : board.whitePieceList != null)
-            return false;
-        return Arrays.deepEquals(tiles, board.tiles);
+        return whitePieceList != null ? whitePieceList.equals(board.whitePieceList) : board.whitePieceList == null;
     }
 
     @Override
     public int hashCode() {
         int result = blackPieceList != null ? blackPieceList.hashCode() : 0;
         result = 31 * result + (whitePieceList != null ? whitePieceList.hashCode() : 0);
-        result = 31 * result + Arrays.deepHashCode(tiles);
         return result;
     }
 }
